+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [FEATURE]
 * [ENHANCEMENT] Include TCP OutRsts in netstat metrics
 * [ENHANCEMENT] Added XFS inode operations to XFS metrics
+* [ENHANCEMENT] Remove CGO dependencies for OpenBSD amd64
 * [BUGFIX]
 
 ## 1.0.1 / 2020-06-15